Updated to Elm 0.19 aliases and functions.
Updates all aliases and functions to Elm 0.19 syntax. The use of `elm
package` has been replaced with `elm install`.

Added the `elmg` alias for initializing new Elm applications (default).
To build a package instead, use `elmg` but edit the `type` to be
`package` in the generated `elm.json` file.

Sadly, the `--warn` option for `elm make` has been removed but there is
talk that it will be restored in a future update. The `--optimize`
option has been added to `elm make` in order to always build an
optimized/compressed version for deployment purposes.