Updated Git global pull config to preserve when rebasing.
Ensures locally committed merge commits are not flattened when running
`git pull`. The use of `pull.rebase = preserve` is virtually the same
as `git rebase --preserve-merges`.

Improves the rebase experience when dealing with *dirty* repositories
that don't adhere to a strict Git Rebase Workflow by allowing the
rebaser to move feature branch commits ahead of the merges while
retaining the merge commits therein. Without this support, the rebaser
would have to sift through the merge commits that would have been
flattened with no context of the merge bubble retained.